refactor: move const.incrementals into profiles.
The incrementals list is basically for profile data, even
if portage_conf and domain was written otherwise.  Thus
move the list into profiles (out of const); this is in
preparation for eventually refactoring the incremental
configuration processing from portage_conf and domain into
a profile subclass so the logic is contained all there.

Signed-off-by: Brian Harring <ferringb@gmail.com>