refactor: loosen prototype.tree._get_* requirements, add typing.
This moves the burden of derivatives having to ensure they return
immutables, instead this is handled by the internals of prototype.tree
since it already partially was.

This change basically makes it easier to write generators for
cat/pkg/ver listing.

For existing implementations returning immutable objects this has
no impact; `t=(1,2);assert t is tuple(t)` for example, and the
same holds for frozenset.

Signed-off-by: Brian Harring <ferringb@gmail.com>